agent/requirements.txt
bedrock-agentcore
strands-agents
strands-agents-tools
boto3>=1.35.0
requests>=2.31.0
